new XML parser
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@5 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ test.sh
@@ -8,8 +8,9 @@ if [ $numberOfdllsNeeded -ne $numberOfdllsExpected ]
 	then echo "you probably don't have all project specific libraries statically linked, please check before BOINC deployment!"
 fi
 
-rm test/sandbox/* 2>/dev/null
-cp model/openMalaria test/sandbox
+rm -rf test/sandbox 2>/dev/null
+mkdir -p test/sandbox
+cp model/openMalaria test/sandbox/openMalaria
 strip test/sandbox/openMalaria
 cd test/sandbox && cp ../original/* . 2>/dev/null
 
@@ -19,8 +20,18 @@ runScenario() {
   ./openMalaria  > /dev/null 
   wait
   date
-  mv output.txt output$number.txt
-  test -e output$number.txt && ../original/compareOutputsFloat.py original$number.txt output$number.txt 1
+  if [ -e stderr.txt ]
+  then
+    mv stderr.txt stderr$number.txt
+  fi
+  if [ -e output.txt ]
+    then
+      mv output.txt output$number.txt
+      ../original/compareOutputsFloat.py original$number.txt output$number.txt 1
+    else
+      echo "No output; error messages:"
+      cat stderr$number.txt
+  fi
 }
 
 if [ "$#" -eq "0" ]
